Move code out of VisitorStateModifications
